Add name argument to optimize, modify search space
The major update the addition of the name argument to the optimize
command line tool. Previously, multiple invocations of optimize would
cause them to crash as they all attempted to create, write, read, and
close the same file (i.e. TMVAClassification.root). Adding a name
prevents this unfortunate error, and on the upside helps the user to
differentiate between the results of different optimization runs.

The ranges of the following hyperparameters have been expanded:

- NTrees
- MaxDepth
- nCuts